+++ DQMServices/Components/plugins/DQMScalInfo.cc
