+++ DQMServices/Components/plugins/DQMScalInfo.h
