+++ DQMServices/Components/plugins/SealModule.cc
@@ -4,33 +4,33 @@
 
 
 // The module providing event information 
-#include "DQMServices/Components/src/DQMEventInfo.h"
+#include "DQMEventInfo.h"
 DEFINE_FWK_MODULE(DQMEventInfo);
 #include "DQMServices/Components/interface/QualityTester.h"
 DEFINE_FWK_MODULE(QualityTester);
-#include "DQMServices/Components/src/DQMFileSaver.h"
+#include "DQMFileSaver.h"
 DEFINE_FWK_MODULE(DQMFileSaver);
-#include "DQMServices/Components/src/DQMFEDIntegrityClient.h"
+#include "DQMFEDIntegrityClient.h"
 DEFINE_FWK_MODULE(DQMFEDIntegrityClient);
-#include "DQMServices/Components/src/DQMStoreStats.h"
+#include "DQMStoreStats.h"
 DEFINE_FWK_MODULE(DQMStoreStats);
-#include "DQMServices/Components/src/DQMMessageLogger.h"
+#include "DQMMessageLogger.h"
 DEFINE_FWK_MODULE(DQMMessageLogger);
-#include "DQMServices/Components/src/DQMMessageLoggerClient.h"
+#include "DQMMessageLoggerClient.h"
 DEFINE_FWK_MODULE(DQMMessageLoggerClient);
-#include "DQMServices/Components/src/DQMFileReader.h"
+#include "DQMFileReader.h"
 DEFINE_FWK_MODULE(DQMFileReader);
-#include "DQMServices/Components/src/DQMProvInfo.h"
+#include "DQMProvInfo.h"
 DEFINE_FWK_MODULE(DQMProvInfo);
-#include "DQMServices/Components/src/DQMDcsInfo.h"
+#include "DQMDcsInfo.h"
 DEFINE_FWK_MODULE(DQMDcsInfo);
-#include "DQMServices/Components/src/DQMDcsInfoClient.h"
+#include "DQMDcsInfoClient.h"
 DEFINE_FWK_MODULE(DQMDcsInfoClient);
-#include "DQMServices/Components/src/DQMScalInfo.h"
+#include "DQMScalInfo.h"
 DEFINE_FWK_MODULE(DQMScalInfo);
 
 // Data Certification module for DAQ info
-#include "DQMServices/Components/src/DQMDaqInfo.h"
+#include "DQMDaqInfo.h"
 DEFINE_FWK_MODULE(DQMDaqInfo);
 
 // module converting between ME and ROOT in Run tree of edm file
